More on `::` briefly in packages.Rmd
Because it is one of the fundamental benefits of the operator, I thought it would be worth putting in a little bit that explains `::` can be used to prevent namespace conflicts. It seems reasonable, because I think novice R users could otherwise believe that using `conflicted` is the only way to work. But this is just a suggestion.